requirements.txt
transformers
datasets
peft
bitsandbytes
Pillow
torch
trl
